+++ packages/authentication/src/jwt.ts
@@ -3,6 +3,8 @@ import { omit } from 'lodash';
 import { IncomingMessage } from 'http';
 import { NotAuthenticated } from '@feathersjs/errors';
 import { Params } from '@feathersjs/feathers';
+// @ts-ignore
+import lt from 'long-timeout';
 
 import { AuthenticationBaseStrategy } from './strategy';
 import { AuthenticationRequest, AuthenticationResult, ConnectionEvent } from './core';
@@ -30,16 +32,19 @@ export class JWTStrategy extends AuthenticationBaseStrategy {
     const isValidLogout = event === 'logout' && connection.authentication && authResult &&
       connection.authentication.accessToken === authResult.accessToken;
 
-    if (authResult && event === 'login') {
-      const { accessToken } = authResult;
+    const { accessToken } = authResult || {};
+
+    if (accessToken && event === 'login') {
+      debug('Adding authentication information to connection');
       const { exp } = await this.authentication.verifyAccessToken(accessToken);
       // The time (in ms) until the token expires
       const duration = (exp * 1000) - new Date().getTime();
       // This may have to be a `logout` event but right now we don't want
       // the whole context object lingering around until the timer is gone
-      const timer = setTimeout(() => this.app.emit('disconnect', connection), duration);
+      const timer = lt.setTimeout(() => this.app.emit('disconnect', connection), duration);
 
       debug(`Registering connection expiration timer for ${duration}ms`);
+      lt.clearTimeout(this.expirationTimers.get(connection));
       this.expirationTimers.set(connection, timer);
 
       debug('Adding authentication information to connection');
@@ -51,7 +56,8 @@ export class JWTStrategy extends AuthenticationBaseStrategy {
       debug('Removing authentication information and expiration timer from connection');
 
       delete connection.authentication;
-      clearTimeout(this.expirationTimers.get(connection));
+      lt.clearTimeout(this.expirationTimers.get(connection));
+      this.expirationTimers.delete(connection);
     }
   }
 
@@ -124,8 +130,7 @@ export class JWTStrategy extends AuthenticationBaseStrategy {
   }
 
   async parse (req: IncomingMessage) {
-    const result = { strategy: this.name };
-    const { header, schemes }: { header: any, schemes: string[] } = this.configuration;
+    const { header, schemes }: { header: string, schemes: string[] } = this.configuration;
     const headerValue = req.headers && req.headers[header.toLowerCase()];
 
     if (!headerValue || typeof headerValue !== 'string') {
@@ -134,7 +139,7 @@ export class JWTStrategy extends AuthenticationBaseStrategy {
 
     debug('Found parsed header value');
 
-    const [ , scheme = null, schemeValue = null ] = headerValue.match(SPLIT_HEADER) || [];
+    const [ , scheme, schemeValue ] = headerValue.match(SPLIT_HEADER) || [];
     const hasScheme = scheme && schemes.some(
       current => new RegExp(current, 'i').test(scheme)
     );
@@ -144,7 +149,7 @@ export class JWTStrategy extends AuthenticationBaseStrategy {
     }
 
     return {
-      ...result,
+      strategy: this.name,
       accessToken: hasScheme ? schemeValue : headerValue
     };
   }